Add better error checking as per #12

+++ VisTarsier.NiftiLib/Processing/Tools.cs
@@ -45,6 +45,25 @@ namespace VisTarsier.NiftiLib.Processing
             // This is in case we have a reference slide at the front or end of the dicom stack, which can happen.
             var nims = outFiles.Where(f => Path.GetExtension(f) == ".nii").OrderByDescending(f => new FileInfo(f)?.Length);
             var nim = nims.FirstOrDefault();
+
+            if(nim == null)
+            {
+                if (!File.Exists(CapiConfig.GetConfig().Binaries.dcm2niix))
+                {
+                    Log.GetLogger().Error("Could not find dcm2niix at: " + CapiConfig.GetConfig().Binaries.dcm2niix);
+                }
+
+                var log = Log.GetLogger();
+                log.Error("Could not find valid output for dcm2niix. Files output were: ");
+
+                foreach (var of in outFiles)
+                {
+                    log.Error($"[{of}]");        
+                }
+
+                throw new FileNotFoundException("Could not find output of dcm2niix");
+            }
+
             if (File.Exists(niftiPath)) File.Delete(niftiPath);
             File.Move(nim, niftiPath);
